rename selectors/event -> selectors/event-deprecated; rename selectors/events -> selectors/events-deprecated;
