+++ src/components/PerformanceList.js
@@ -9,7 +9,7 @@ import LayoutColumn from "./LayoutColumn";
 import Text from "./Text";
 import SectionDivider from "./SectionDivider";
 
-import { getTimePeriod } from "../selectors/events";
+import { getTimePeriod } from "../selectors/events-deprecated";
 import type { Performance, PerformancePeriods } from "../data/event-deprecated";
 import { sectionHeaderShadow, sectionHeaderBgColor } from "../constants/colors";
 import text from "../constants/text";
