+++ src/screens/CategoriesFilterScreen/index.js
@@ -5,7 +5,7 @@ import type { NavigationScreenProp, NavigationState } from "react-navigation";
 import type { State } from "../../reducers";
 import type { Event, EventCategoryName } from "../../data/event-deprecated";
 import { setEventFilters } from "../../actions/event-filters";
-import { selectFilteredEvents } from "../../selectors/events";
+import { selectFilteredEvents } from "../../selectors/events-deprecated";
 import Component from "./component";
 
 type OwnProps = {
