+++ src/selectors/basic-event-filters.js
@@ -6,7 +6,7 @@ import {
   getHours,
   diff as dateDiff
 } from "../lib/date";
-import { selectEventIsFree } from "./event";
+import { selectEventIsFree } from "./event-deprecated";
 import areaBoundaries from "../data/areas";
 import type { Event, EventCategoryName } from "../data/event-deprecated";
 import type { DateRange, Time } from "../data/date-time";
